Added RankBoostTools and refractored.
